Backend/SearchByImageAndOCRPDF/LawFlow_SearchByImage/requirements.txt
Flask==2.2.1
Flask-RESTful==0.3.9
gunicorn==20.1.0
bcrypt==3.2.2
blueprint==3.4.2
pandas==1.3.5
PyPDF2==2.10.5
pydub==0.25.1
tokenlib==2.0.0
google-cloud==0.34.0
firebase==3.0.1
firebase-admin==5.2.0
google-cloud-vision==3.1.1
google-cloud-storage==2.4.0
oauthlib==3.2.1
python-jwt==3.3.3
python-dotenv==0.21.0
path==16.4.0
# pandas==1.4.3
PyJWT==2.4.0
uuid==1.30
Flask-Cors==3.0.10
sklearn==0.0
scikit-learn==1.0.2
nltk==3.7 
networkx==2.6.3
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
scipy==1.9.2
#en-core-web-sm==3.4.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl